% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_helpers.R
\name{fread_pcv}
\alias{fread_pcv}
\title{Read a PCV file}
\usage{
fread_pcv(...)
}
\arguments{
\item{...}{Other options passed on to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
Returns a PCV data table read using \code{\link[data.table:fread]{data.table::fread()}} function call.
}
\description{
This function reads a PCV (Primary Care Visits) table written as a CSV file using \code{\link[data.table:fread]{data.table::fread()}} and allows additional
parameters to be passed to \code{fread()}.
}
\seealso{
\code{\link[data.table:fread]{data.table::fread()}} for the underlying functionality.
}
