% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_helpers.R
\name{fread_sip}
\alias{fread_sip}
\title{Read a SIP file}
\usage{
fread_sip(...)
}
\arguments{
\item{...}{Other options passed to \code{\link[data.table:fread]{data.table::fread()}}}
}
\value{
a \code{\link[data.table:fread]{data.table::fread()}} function call
}
\description{
Read a SIP file
}
